Generar pdf

+++ database/migrations/2025_02_21_171848_add_consumido_to_dieta_alimentos_table.php
@@ -1,21 +1,25 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration {
     public function up()
     {
         Schema::table('dieta_alimentos', function (Blueprint $table) {
-            $table->boolean('consumido')->default(false)->after('cantidad'); // Añadir la columna después de "cantidad"
+            if (!Schema::hasColumn('dieta_alimentos', 'consumido')) {
+                $table->boolean('consumido')->default(0)->after('cantidad');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('dieta_alimentos', function (Blueprint $table) {
-            $table->dropColumn('consumido'); // Eliminar la columna si se revierte la migración
+            if (Schema::hasColumn('dieta_alimentos', 'consumido')) {
+                $table->dropColumn('consumido');
+            }
         });
     }
 };
